Moving preprocessing into its own file as it was getting too big
