+++ emo_eval/__init__.py
@@ -1,0 +1,2 @@
+from .preprocessing import preprocess
+from .emo_eval import main, feature_extraction, train_crossval
