+++ main.py
@@ -1,0 +1,3 @@
+from emo_eval import main
+
+main("data/train.txt")
